LPS-44538 move from PrincipalThreadLocal to PasswordModificationThreadLocal
