+++ portal-service/src/com/liferay/portal/kernel/security/auth/PrincipalThreadLocal.java
@@ -18,7 +18,6 @@ import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.util.AutoResetThreadLocal;
 import com.liferay.portal.kernel.util.GetterUtil;
-import com.liferay.portal.kernel.util.StringPool;
 
 /**
  * @author Brian Wing Shun Chan
@@ -39,16 +38,8 @@ public class PrincipalThreadLocal {
 		return _password.get();
 	}
 
-	public static String getPasswordUnencrypted() {
-		return _passwordUnencrypted.get();
-	}
-
 	public static long getUserId() {
 		return GetterUtil.getLong(getName());
-	}
-
-	public static boolean isPasswordModified() {
-		return _passwordModified.get();
 	}
 
 	public static void setName(long name) {
@@ -67,14 +58,6 @@ public class PrincipalThreadLocal {
 		_password.set(password);
 	}
 
-	public static void setPasswordModified(boolean passwordModified) {
-		_passwordModified.set(passwordModified);
-	}
-
-	public static void setPasswordUnencrypted(String passwordUnencrypted) {
-		_passwordUnencrypted.set(passwordUnencrypted);
-	}
-
 	private static final Log _log = LogFactoryUtil.getLog(
 		PrincipalThreadLocal.class);
 
@@ -82,12 +65,5 @@ public class PrincipalThreadLocal {
 		PrincipalThreadLocal.class + "._name");
 	private static final ThreadLocal<String> _password =
 		new AutoResetThreadLocal<>(PrincipalThreadLocal.class + "._password");
-	private static final ThreadLocal<Boolean> _passwordModified =
-		new AutoResetThreadLocal<>(
-			PrincipalThreadLocal.class + "._passwordModified", false);
-	private static final ThreadLocal<String> _passwordUnencrypted =
-		new AutoResetThreadLocal<>(
-			PrincipalThreadLocal.class + "._passwordUnencrypted",
-			StringPool.BLANK);
 
 }
